+++ noodles-csi/src/io.rs
@@ -1,5 +1,6 @@
 //! CSI I/O.
 
+mod filtered_lines;
 mod query;
 
-pub use self::query::Query;
+pub use self::{filtered_lines::FilteredLines, query::Query};